Added MetadataDocument type
